eliminated encryption package in favor of non-internal 'temporalutil' package, then refactored code accordingly. I also moved the encrypt and decrypt functions inside the data converter.
