+++ app/temporalutil/encryption_test.go
@@ -1,4 +1,4 @@
-package encryption
+package temporalutil
 
 import (
 	"testing"
